added temperature list to the createRoomBoundary component.
temperatureController and Dbservices updated to return temperature id.

+++ backend/models/roomBoundaryModel.js
@@ -56,7 +56,7 @@ const roomBoundary = (sequelize, DataTypes) => {
         defaultValue: null,
       },
       area: {
-        type: DataTypes.DECIMAL(10, 3),
+        type: DataTypes.FLOAT,
         required: true,
         notNull: true,
         notEmpty: true,
